requirements.txt
pip==20.2.3
setuptools==50.3.0
wheel==0.35.1
pycparser==2.20
earl-etf==2.1.2
PyYAML==5.3.1
numpy==1.19.2
wsaccel==0.6.2
-e git+https://github.com/elderlabs/BetterDisco.git@v8/py3-testing#egg=disco-py
click==7.1.2
Flask==1.1.2
greenlet==0.4.17
-e git+https://github.com/elderlabs/holster#egg=holster
peewee==3.13.3
psycogreen==1.0.2
psycopg2-binary==2.8.6
redis==3.5.3
werkzeug==1.0.1
emoji==0.6.0
fuzzywuzzy==0.18.0
GitPython==3.1.9
humanize==3.0.1
markovify==0.8.3
regex==2020.9.27
sentry-sdk[flask]==0.18.0
pytz==2020.1
Unidecode==1.1.1
cffi==1.14.3
cairocffi==1.1.0
Pillow==7.2.0
CairoSVG==2.4.2
-e git+https://github.com/elderlabs/pygal#egg=pygal
